// src/model/ActorPermissions.ts
/**
 * ActorPermissions.ts
 * An enum containing the different permissions that an actor can have for the DMP Registry system.
 */

export enum ActorPermissions {
    READ_AGENTS = "READ_AGENTS",
    WRITE_AGENTS = "WRITE_AGENTS",
    READ_REGISTRY = "READ_REGISTRY",
    WRITE_REGISTRY = "WRITE_REGISTRY",
    MANAGE_USERS = "MANAGE_USERS"
}